refactor(tenants): enhance DTOs with proper validation and TypeScript types
