+++ backend/src/modules/tenants/dto/create-tenant.dto.ts
@@ -17,7 +17,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { AddressDto } from '../../../common/dto/address.dto';
 import { ContactInfoDto } from '../../../common/dto/contact-info.dto';
-import { BusinessType, BusinessScale } from '../entities/tenant.entity';
+import { BusinessType, BusinessScale } from '../enums/tenant.enums';
 
 export class CreateTenantDto {
   @ApiProperty({
